remove unsolved

+++ src/datasets.py
@@ -1112,7 +1112,7 @@ class AlgebraAIDataset(torch.utils.data.Dataset):
             dataset['problems'] = d['problems']
         res = np.array(dataset['response'], dtype=int)
         num_correct = res.sum(0)
-        unsolved_idx = num_correct != 0
+        unsolved_idx = (num_correct != 0).astype(int)
         if is_train:
             self.n_students = len(dataset['epoch'])
             self.n_problems = len(dataset['problems'][unsolved_idx])
